Only attempt proxyInitting if __init__ has been found and worked, and only actually init instances of PyProxy.
Add an __init__ for all Java classes, even interfaces and abstract classes.
Handle throwing type errors for uninstantiable classes in
PyReflectedConstructor.

Pass _is and _isnot through to the javaproxy == if javaproxy is around.  This
assumes a constant mapping between a given Java type and the PyType wrapping
it, but I think that's safe.