Undoing a change that got checked in here
http://fisheye3.atlassian.com/changelog/jython/?cs=4791

Which came originally from here

http://svn.python.org/view/python/trunk/Lib/socket.py?view=diff&r1=61007&r2=61008

But was wrong; changing min() to max() solves this problem.

The cpython change was ostensibly made to correct a memory allocation problem, but was wrong, and was undone here

http://svn.python.org/view?view=rev&revision=62627

We should probably review _fileobject implementation sometime, since the cpython one has now changed a fair amount.