Restore select.cpython_compatible_select
Jython 2.5 has select.cpython_compatible_select that -- as its name
suggest -- is compatible with how select.select behaves in CPython
with respect to blocking sockets. Now that select.select has this
desired compatible behavior, without any special workarounds, this
function is obsolete. However, there is code for Jython that expects
it, so now have it be an alias.