os.getenv, os.listdir may return unicode instead of bytes
Note that now this deviates from CPython 2.7, although matches the
behavior seen in Python 3.x. Here's why this is necessary:

* Java does not support paths and environment entries (variables or
  values) not expressible in Unicode. Similar restrictions are seen in
  CPython 2.7: OS X does not support non Unicode directory names.

* However Jython's raison d'?tre is seamless Java integration: it
  should be generally possible to mix Python and Java together, eg
  [java.io.File(p).exists() for p in os.listdir()] should always work.

In general, this changeset should not have impact on user code, given
that for code using Python 2, unicode and str/bytes can be freely
mixed. For Jython specifically, os.listdir and functionality that
builds on it like glob has been until now broken for Unicode paths
(with codepoints > 255) since Jython 2.5.

Fixes an urgent bug that Jython could not be started with a current
working directory with a path using Unicode codepoints > 255, part of
http://bugs.jython.org/issue2239

Completes an outstanding issue for OSError and Unicode paths, see
http://bugs.jython.org/issue1825