Updated ASM to 5.0.3 for Java 8 support
Using ASM 4 to target Java 8 works for code generation, but not for
code introspection, such as analyzing jars or supporting Clamp. Fixed
by upgrading to ASM 5. Replaced Jython-specific code to analyze access
permissions from a classfile in favor of ASM.

For Jython developers working on code compilation: when emitting
bytecode, if you are visiting a method that's owned by an interface
(not implemented by a class), you need to use different
visitMethodInsn. See the change in CodeCompiler#visitWith - the only
change that was currently required.

Bumped bytecode magic.