jython -m test.regrtest -e now finds the tests it should. Fixes #2332.
Aligns findtestdir() with later lib-python version of regrtest so it
finds the tests without reliance on sys.argv[0]. Also adds a method
that legitimizes use #-comments in the tables of expected skips and
failures. (We already do, but they were appearing as test names.) The
logic supporting the --expected flag is clarified.