Avoid deadlock in running closers. Fixes #2457
This is a further elaboration of the work done in #2280, and should be
completely compatible (and has been tested) with the recent change in
r7889:cddadb6808c0.

Jim's theory (and Jim is the committer of this change on Nick's
behalf): synchronizing on the class, vs an object, is a bad idea
because of class unloading that's happening at shutdown. At the very
least this is now tied to the closer itself.