Refactor array-based PyBuffer implementations.
Quite substantial change to the class hierarchy for the array-based
buffers to permit sharing with a parallel one using ByteBuffer. The
ByteBuffer implementation is not in this commit (see later). There is
a bug in copyFrom, when copying between buffers exported by the same
object, that causes test_memoryview to fail.