Fix all present irregularities in the javadoc
It seems Gradle will not publish defective javadoc, so it has to be correct.
Most problems were due to imaginative choice of argument in @see and @link tags.

No code changes except that the solution was sometimes to add an import
statement for something referred to in documentation. The editor imposes a few
coding rules on save (trailing space, @Override, parentheses) but I have not
reformatted much, even code sorely divergent from our standard.