Fix star-import on Java 9 for selected modules, partially addresses #2362.
CachedJarsPackageManager now creates PyJavaPackage objects with class lists,
for packages in modules, as we already do for JARs. This is used, when the
"jrt:" filesystem exists (Java 9), from SysPackageManager::findAllPackages().
Some refactoring and a lot of new comment is included, for intelligibility in
this and further work.

No cache file is created at present, and we process only a configured list of
modules, both shortcomings we may address in further work.

An update of asm to 7.0 is necessary to parse the accessibility information
from Java runtime classes without instantiating them.