Remove actual clone operation from build.xml and refactor.
This removal simplifies the structure of the build (no work.dir). The
consequent risk of releasing from a polluted baseline is mitgated by
the target force-snapshot-if-polluted. It is now easier to build an
installer intended as a snapshot release.

Parts of build.xml have been heavily restructured in order to sequence
the logic behind this idea, probably improving them. Resource type
filelist/file replaces fileset/path for specifying JARs.

Some rough edges remain but it's already a lot for one change set.