Create a __main__ module and set __name__ in JSR-223 (fixes #2846).
This change causes the default engine context in a PyScriptEngine to be
created as a module __main__, and __name__ to be defined, comparable to
the environment found when executing a Python program. The thread-local
name space concept is otherwise undisturbed.

org.python.jsr223.ScriptEngineTest cases are modified correspondingly,
and testThreadLocalBindings is fixed to test isolation within a thread.